fix(aq5): CI lint gate fixes — RULE-002/003 splits, stale version pin, hermetic pytests
- crates/atm-daemon-bootstrap/src/queue_drain.rs: split drain_one (120-line
  RULE-002 hard violation) into claim_drain_target/emit_drain_dispatch/
  clear_drain_handoff_marker helpers, each owning one stage of the claim ->
  emit -> marker-clear pipeline.
- crates/atm-daemon-bootstrap/src/lib.rs: extract the peer-wire-mode,
  direct-peer-port, and peer-pool launch/env parsing block into a new
  peer_wire_config module (RULE-003 file-size gate; lib.rs prod lines
  1080 -> under 1000). Public API re-exported unchanged; architecture
  guard call-site checks (ao2/ao4) are unaffected since they scan call
  sites, not the parsed definitions.
- crates/atm-core/Cargo.toml: fix a stale 1.4.4 dev-dependency version pin
  on atm-storage left over from a rebase base, which failed the version/
  manifests lint gates after merging in AQ1.6's later commits.
- crates/atm-architecture/tests/graft_receiver_ownership_boundary.rs: point
  the RegisteredGraftReceiver Drop/unregister structural guard at
  runtime/lease_client.rs, where AQ1.6 QA-2's RULE-003 split moved that
  code (runtime.rs no longer exists at the old path).
- .just/tests/test_send_to_surface.py + scripts/send-to/atm-send-to.{sh,ps1}:
  root-caused the wyvern-degradation pytest failures — every degradation
  case (absent/below/unparsable/hang/missing-asset/unknown-schema) falls
  through to the OS-native picker, which requires zenity/fzf/osascript to
  actually be installed on the runner. Added a test-only
  ATM_SEND_TO_NATIVE_PICKER override (mirrors the existing
  ATM_SEND_TO_PICKER/WYVERN_BIN/WYVERN_ASSET seams in both the sh and ps1
  pipelines) so the harness never touches real desktop tooling, and gave
  the "hang" fixture a much larger sleep margin over probe_wyvern.py's
  1.5s bounded deadline. atm-send-to.sh is the macOS/Linux pipeline only —
  Windows already ships a separate atm-send-to.ps1 + picker-windows.ps1
  pipeline validated by manual E2E (sprint-AQ5-surface-evidence.md); the
  pytest class is now skipped on win32 to match the existing
  test_release_gate.py precedent for the same reason, instead of pretending
  the bash/zenity/fzf script has a Windows target it was never designed for.

Verified: cargo fmt --all, cargo clippy --workspace --exclude atm-daemon
--all-targets -- -D warnings, cargo test --workspace --exclude atm-daemon
(all green), python3 .just/run_lint.py all (32/32 checks pass, including
sc-boundary and pytests).
